requirements.txt
Django==1.11.13
djangorestframework==3.8.2
pytz==2018.4
pylint==1.9.2
pylint-django==0.11.1
python-dateutil==1.5
